Doing more work on the I/O module
The idea behind this design is to capsule all file operations away
into a single trait that can easily be implemented (with no functions).

Then there are both files and metadata files. A metadata file also
implements the trait `MetadataFile<T: Storable>` which can fetch all
storables in it's domain, effectively filling it's parent with data.

Maybe we could replace the fetch returning a Vec<T> with an Iterator<Item=T>
so that we can do the loading more lazily or easily with rayon in the
future? Whoooo knows... 🤷