Adding & Refactoring file management for Chapters & Scenes
Basically: Reworking some of the traits to make it easier to pull
in files and metadata files.
This really isn't a perfect way of doing this but it's the best we
can do right now. We should keep in mind that we want to expand to
more complicated file relationships (Chapter/Chapter/File, etc) and
then this layout will _not_ be enough.

But it's enough for now :)