Enforcing a one-to-one mapping of Gtk to Novelist types.
This will ensure that each component has a primary type
it exports onto a higher interface (i.e. `gtk::Window`)
while also allowing it to have private "auxilliary" types
that are shadowed and abstracted.

Adding `as_ref()` and `as_mut_ref()` as functions to all
Components that allow the retrieval of these wrapped types.
This should be much nicer to use than `Deref` because it's
not overloading the `*` operator or using implicit derefs.

We might want to be able to generate the "getters" with
some macro at some point...